api/pyproject.toml
[tool.poetry]
name = "comses-rag"
version = "0.1.0"
description = ""
authors = ["Anton Suharev <asuharev@asu.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.7"
fastapi = "^0.115.4"
langchain-community = "^0.3.5"
langchain-core = "^0.3.15"
sqlalchemy = "^2.0.35"
langfuse = "^2.53.9"
pydantic = "^2.9.2"
bcrypt = "^4.2.0"
uvicorn = "^0.32.0"
pgvector = "^0.3.6"
python-multipart = "^0.0.17"
pyjwt = "^2.9.0"
psycopg2-binary = "^2.9.10"
temporalio = "^1.8.0"
langchain-ollama = "^0.2.0"
openai = "^1.54.4"
instructor = "^1.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
